feat(zeebe): add zero conf and docs
