+++ README.md
@@ -27,6 +27,7 @@ This package can be useful because:
 
 -   [Documentation is available in this folder](packages/workit-camunda/.docs/)
 -   Comprehensive API documentation is available [online](https://villedemontreal.github.io/workit/) and in the `docs` subdirectory
+-   [Examples](examples)
 
 ## Packages
