+++ packages/workit-camunda/src/config/container.ts
@@ -6,7 +6,6 @@
 
 import { camundaLogger, logger } from 'workit-bpm-client';
 import { kernel } from 'workit-core';
-import { ICamundaConfig } from 'workit-types';
 import { constants } from './constants';
 import { SERVICE_IDENTIFIER } from './constants/identifiers';
 
@@ -20,27 +19,3 @@ Object.values(constants.envs).forEach(env => {
     .toConstantValue(logger)
     .whenTargetNamed(env);
 });
-
-if (!process.env.SKIP_DEMO_CONFIG) {
-  const configBase: ICamundaConfig = {
-    workerId: 'demo',
-    baseUrl: `__undefined__`,
-    topicName: 'topic_demo'
-  };
-
-  const bpmnPlatformClientConfig = Object.assign({}, configBase, {
-    baseUrl: `http://localhost:8080/engine-rest`,
-    maxTasks: 32,
-    autoPoll: false
-  });
-
-  const zeebeElasticExporterConfig = {
-    url: `http://localhost:9200`
-  };
-
-  const zeebeClientConfig = Object.assign({}, configBase, { baseUrl: `localhost:26500` });
-
-  kernel.bind(SERVICE_IDENTIFIER.camunda_external_config).toConstantValue(bpmnPlatformClientConfig);
-  kernel.bind(SERVICE_IDENTIFIER.zeebe_external_config).toConstantValue(zeebeClientConfig);
-  kernel.bind(SERVICE_IDENTIFIER.zeebe_elastic_exporter_config).toConstantValue(zeebeElasticExporterConfig);
-}
